# api-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "description": "A simple Hyper-Express API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/realm": {
      "post": {
        "summary": "Create a new realm",
        "tags": [
          "Realm"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "realmName": {
                    "type": "string",
                    "example": "example.domain.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Realm created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Realm created"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/test-secure": {
      "post": {
        "summary": "Test secure route",
        "description": "This endpoint is protected and requires both a valid JWT token in the Authorization header.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "User has the right to access this route.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "user has right"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access due to missing, invalid, or expired token, or incorrect password.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "summary": "Log in a user",
        "description": "Authenticates a user with their username, password, and realmName.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "user123"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  },
                  "realmName": {
                    "type": "string",
                    "example": "exampleRealm"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "refreshToken": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Validation error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "post": {
        "summary": "Create a new user",
        "description": "Endpoint to create a new user.",
        "tags": [
          "User"
        ],
        "operationId": "createUser",
        "requestBody": {
          "description": "Request payload for creating a new user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "realmName": {
                    "type": "string",
                    "description": "The name of the user's realm"
                  },
                  "username": {
                    "type": "string",
                    "description": "The user's username"
                  },
                  "firstName": {
                    "type": "string",
                    "description": "The user's first name"
                  },
                  "lastName": {
                    "type": "string",
                    "description": "The user's last name"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The user's email address"
                  },
                  "password": {
                    "type": "string",
                    "description": "The user's password"
                  }
                },
                "required": [
                  "realmName",
                  "username",
                  "firstName",
                  "lastName",
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User created"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "examples": {
                        "validationError": {
                          "value": "validation Error"
                        },
                        "otherError": {
                          "value": "Error message"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}